Add files via upload
In number theory, a number n is called perfect if the sum of its proper divisors is equal to n, 
deficient if the sum of its proper divisors is less than n, and abundant if the sum of its proper 
divisors exceeds n. Write a Web-based Django application to perform the following: [Note: No 
necessity to create Model or Form]
i. Write a function in views.py that returns the list of proper divisors for a given number 
[Note: Proper divisor means all the divisors of that number except that number itself. For 
eg. Proper divisors of 10 are: 1, 2, and 5]
ii. Write a function in views.py that generates and returns the list of perfect, deficient and 
abundant numbers within a given range using list comprehension 
iii. Create a webpage using template systems to render and print the result