Add files via upload
 In number theory, Amicable numbers are two different numbers so related that the sum of the 
proper divisors of each is equal to the other number. The smallest pair of amicable numbers is 
(220, 284). They are amicable because the proper divisors of 220 are 1, 2, 4, 5, 10, 11, 20, 22, 44, 
55 and 110, of which the sum is 284; and the proper divisors of 284 are 1, 2, 4, 71 and 142, of 
which the sum is 220. Write a Python program to compute the sum of all the amicable numbers 
under a given number. Write a Web-based Django application to perform the following: [Note: 
No necessity to create Model or Form]
i. Write a function that returns the list of proper divisors for a given number [Note: Proper 
divisor means all the divisors of that number except that number itself. For eg. Proper 
divisors of 10 are: 1, 2, and 5]
ii. Write a function that generates and returns the list of perfect, deficient and abundant 
numbers within a given range using list comprehension 
iii. Write a function that takes two numbers as parameters and checks and returns whether 
they are amicable or not.
iv. Create a webpage using template systems to render and print whether the following list of 
pairs given as tuples are amicable or not using the function.
 [(150, 290), (220, 284), (1184, 1210), (1350, 1851)]